Separate node name for module name segments
Consider the type `Dict.Dict Int String`. Both `Dict` parts of that type
used to be represented by the name "upper_case_identifier". With this
change that will remain true for the second `Dict`, but the fist `Dict`
will be represented by a node `module_name_segment`.

The motivation for this change is to make it easier to distinguish type
and constructor names from module names.